Split PayloadGeneratorTests into multiple files
